Include the k6 version in the tar archives' metadata.json
