+++ cmd/login_cloud.go
@@ -25,6 +25,7 @@ import (
 
 	"gopkg.in/guregu/null.v3"
 
+	"github.com/loadimpact/k6/lib/consts"
 	"github.com/loadimpact/k6/stats/cloud"
 	"github.com/loadimpact/k6/ui"
 	"github.com/pkg/errors"
@@ -95,7 +96,7 @@ This will set the default token used when just "k6 run -o cloud" is passed.`,
 			email := vals["Email"].(string)
 			password := vals["Password"].(string)
 
-			client := cloud.NewClient("", k6Conf.Collectors.Cloud.Host.String, Version)
+			client := cloud.NewClient("", k6Conf.Collectors.Cloud.Host.String, consts.Version)
 			res, err := client.Login(email, password)
 			if err != nil {
 				return err
